Fix certmonitor labelselector (#1998)
* Fix certMonitor labelSelector

* Fix certMonitor labelSelector

+++ internal/certmonitor/certmonitor.go
@@ -2,14 +2,14 @@ package certmonitor
 
 import (
 	"crypto/x509"
-	"encoding/base64"
 	"encoding/pem"
 	"fmt"
-	"github.com/golang/glog"
 	"github.com/pkg/errors"
 	"github.com/stackrox/acs-fleet-manager/fleetshard/pkg/fleetshardmetrics"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
+	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/validation/field"
 	"k8s.io/client-go/informers"
 	"k8s.io/client-go/tools/cache"
@@ -55,15 +55,13 @@ func (c *certMonitor) Start() error {
 		return errors.New("already started")
 	}
 	c.stopCh = make(chan struct{})
-
 	c.secretInformer.AddEventHandler(cache.ResourceEventHandlerFuncs{
 		AddFunc:    c.handleSecretCreation,
 		UpdateFunc: c.handleSecretUpdate,
 		DeleteFunc: c.handleSecretDeletion,
 	})
 	c.informerfactory.Start(c.stopCh)
-
-	if !cache.WaitForCacheSync(c.stopCh, c.secretInformer.HasSynced) {
+	if !cache.WaitForCacheSync(c.stopCh) {
 		return fmt.Errorf("timed out waiting for caches to sync")
 	}
 	return nil
@@ -91,20 +89,17 @@ func NewCertMonitor(config *Config, informerFactory informers.SharedInformerFact
 // processSecret extracts, decodes, parses certificates from a secret, and populates prometheus metrics
 func (c *certMonitor) processSecret(secret *corev1.Secret) {
 	for dataKey, dataCert := range secret.Data {
-		certConv, err := base64.StdEncoding.DecodeString(string(dataCert))
-		if err != nil {
-			continue
-		}
 
-		pparse, _ := pem.Decode(certConv)
+		pparse, _ := pem.Decode(dataCert)
 		if pparse == nil {
 			continue
 		}
 
 		certss, err := x509.ParseCertificate(pparse.Bytes)
 		if err != nil {
-			glog.Errorf("Failed to parse certificate %s: %v", dataKey, err)
+			continue
 		}
+
 		expiryTime := float64(certss.NotAfter.Unix())
 		c.metrics.SetCertKeyExpiryMetric(secret.Namespace, secret.Name, dataKey, expiryTime)
 	}
@@ -114,6 +109,9 @@ func (c *certMonitor) processSecret(secret *corev1.Secret) {
 func (c *certMonitor) handleSecretCreation(obj interface{}) {
 	secret, ok := obj.(*corev1.Secret)
 	if !ok {
+		return
+	}
+	if !c.shouldProcessSecret(secret) {
 		return
 	}
 	c.processSecret(secret)
@@ -140,6 +138,11 @@ func (c *certMonitor) handleSecretUpdate(oldObj, newObj interface{}) {
 			c.metrics.DeleteKeyCertMetric(newSecret.Namespace, newSecret.Name, oldKey)
 		}
 	}
+
+	if !c.shouldProcessSecret(newSecret) {
+		return
+	}
+
 	c.processSecret(newSecret)
 }
 
@@ -149,7 +152,65 @@ func (c *certMonitor) handleSecretDeletion(obj interface{}) {
 	if !ok {
 		return
 	}
+	if !c.shouldProcessSecret(secret) {
+		return
+	}
 	c.metrics.DeleteCertMetric(secret.Namespace, secret.Name)
+}
+
+func (c *certMonitor) shouldProcessSecret(s *corev1.Secret) bool {
+	for _, monitor := range c.config.Monitors {
+		if c.secretMatches(s, monitor) {
+			return true
+		}
+	}
+	return false
+}
+
+// secretMatches checks if a secret matches a monitor config
+func (c *certMonitor) secretMatches(s *corev1.Secret, monitor MonitorConfig) bool {
+	if s == nil {
+		return false
+	}
+	if len(monitor.Secret.Name) > 0 && s.Name != monitor.Secret.Name {
+		return false
+	}
+	if len(monitor.Namespace.Name) > 0 && s.Namespace != monitor.Namespace.Name {
+		return false
+	}
+	if monitor.Secret.LabelSelector != nil && !objectMatchesSelector(s, monitor.Secret.LabelSelector) {
+		return false
+	}
+
+	if monitor.Namespace.LabelSelector != nil {
+		ns, err := c.namespaceGetter.Get(s.Namespace)
+		if err != nil {
+			return false
+		}
+		if !objectMatchesSelector(ns, monitor.Namespace.LabelSelector) {
+			return false
+		}
+	}
+	return true
+}
+
+// objectMatchesSelector checks if object matches given label selector
+func objectMatchesSelector(obj runtime.Object, selector *metav1.LabelSelector) bool {
+	if selector == nil {
+		return true
+	}
+	labelselector, err := metav1.LabelSelectorAsSelector(selector)
+	if err != nil {
+		return false
+	}
+
+	metaObj, ok := obj.(metav1.Object)
+	if !ok {
+		return false
+	}
+
+	return labelselector.Matches(labels.Set(metaObj.GetLabels()))
+
 }
 
 // ValidateConfig checks the validity of Config
